pyproject.toml
[tool.poetry]
name = "blame-luca-discord-bot"
version = "0.1.1"
description = "#blameluca tracker for Durham K-pop Society Discord Server"
authors = ["Luca <jgcw74@durham.ac.uk>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.1"
"discord.py" = { git = "https://github.com/Rapptz/discord.py" }

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
